+++ x/sudo/cli/cli.go
@@ -1,9 +1,13 @@
 package cli
 
 import (
+	"encoding/json"
 	"fmt"
 	"os"
 	"strings"
+
+	wasm "github.com/CosmWasm/wasmd/x/wasm/types"
+	"github.com/MakeNowJust/heredoc/v2"
 
 	"github.com/NibiruChain/nibiru/v2/x/sudo"
 
@@ -28,6 +32,7 @@ func GetTxCmd() *cobra.Command {
 	// Add subcommands
 	txCmd.AddCommand(
 		CmdEditSudoers(),
+		CmdEditZeroGasActors(),
 		CmdChangeRoot(),
 	)
 
@@ -60,24 +65,22 @@ func GetQueryCmd() *cobra.Command {
 // function of the sdk.Msg handler for x/sudo.
 func CmdEditSudoers() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "edit [edit-json]",
+		Use:   "edit-sudoers [edit-json]",
 		Args:  cobra.ExactArgs(1),
 		Short: "Edit the x/sudo state (sudoers) by adding or removing contracts",
-		Example: strings.TrimSpace(fmt.Sprintf(`
-			Example: 
-			$ %s tx sudo edit <path/to/edit.json> --from=<key_or_address> 
-			`, version.AppName)),
-		Long: strings.TrimSpace(
-			`Adds or removes contracts from the x/sudo state, giving the 
-			contracts permissioned access to certain bindings in x/wasm.
+		Example: heredoc.Docf(`
+%s tx sudo edit-sudoers <path/to/edit.json> --from=<key_or_address>`, version.AppName),
+		Long: heredoc.Doc(`
+Adds or removes contracts from the x/sudo state, giving the 
+contracts permissioned access to certain bindings in x/wasm.
 
-			The edit.json for 'EditSudoers' is of the form:
-			{
-			  "action": "add_contracts",
-			  "contracts": "..."
-			}
+The edit.json for 'EditSudoers' is of the form:
+{
+  "action": "add_contracts",
+  "contracts": "..."
+}
 
-			- Valid action types: "add_contracts", "remove_contracts"	
+- Valid action types: "add_contracts", "remove_contracts"	
 			`),
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
 			clientCtx, err := client.GetClientTxContext(cmd)
@@ -120,9 +123,8 @@ func CmdChangeRoot() *cobra.Command {
 		Args:  cobra.ExactArgs(1),
 		Short: "Change the root address of the x/sudo state",
 		Example: strings.TrimSpace(fmt.Sprintf(`
-			Example: 
-			$ %s tx sudo change-root <new-root-address> --from=<key_or_address>
-			`, version.AppName)),
+%s tx sudo change-root <new-root-address> --from=<key_or_address>`,
+			version.AppName)),
 		Long: strings.TrimSpace(
 			`Change the root address of the x/sudo state, giving the
 				new address, should be executed by the current root address.
@@ -133,19 +135,69 @@ func CmdChangeRoot() *cobra.Command {
 				return err
 			}
 
+			// marshals contents into the proto.Message (msg)
 			msg := new(sudo.MsgChangeRoot)
-
-			// marshals contents into the proto.Message to which 'msg' points.
 			root := args[0]
+			from := clientCtx.GetFromAddress() // Parse the message sender
+			msg.Sender = from.String()
+			msg.NewRoot = root
+
+			if err = msg.ValidateBasic(); err != nil {
+				return err
+			}
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+// CmdEditZeroGasActors is a terminal command that broadcasts a
+// "nibiru.sudo.v1.MsgEditZeroGasActors" transaction.
+func CmdEditZeroGasActors() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "edit-zero-gas [actors_json_string]",
+		Args:  cobra.ExactArgs(1),
+		Short: "Change the zero gas actors of the x/sudo state",
+		Example: heredoc.Docf(`
+%s tx sudo edit-zero-gas [actors_json_string] --from=<key_or_address>
+
+The [actors_json_string] for "ZeroGasActors" is of the form:
+{
+  "senders": ["nibi1...", "nibi1...", ... ],
+  "contracts": ["0x...", "nibi1....", ... ]
+}
+`, version.AppName),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err
 			}
 
+			jsonBz := []byte(args[0])
+			// JSON validation
+			rawBz := wasm.RawContractMessage(jsonBz)
+			err = rawBz.ValidateBasic()
+			if err != nil {
+				return fmt.Errorf(`invalid arg "%s" is not a JSON string: %w`, jsonBz, err)
+			}
+
+			var zeroGasActors sudo.ZeroGasActors
+			err = json.Unmarshal(jsonBz, &zeroGasActors)
+			if err != nil {
+				return fmt.Errorf("failed to unpack actors json string: %w", err)
+			}
+
 			// Parse the message sender
 			from := clientCtx.GetFromAddress()
-			msg.Sender = from.String()
-			msg.NewRoot = root
 
+			msg := &sudo.MsgEditZeroGasActors{
+				Actors: zeroGasActors,
+				Sender: from.String(),
+			}
 			if err = msg.ValidateBasic(); err != nil {
 				return err
 			}